mv pairwise distance util to array_ops
