+++ mlx_graphs/utils/array_ops.py
@@ -73,3 +73,22 @@ def one_hot(labels: mx.array, num_classes: Optional[int] = None) -> mx.array:
 
     one_hot[mx.arange(shape[0]), labels.squeeze()] = 1
     return one_hot
+
+
+def pairwise_distances(x: mx.array, y: mx.array) -> mx.array:
+    """
+    Compute pairwise distances between points in vectors x and y.
+
+    Args:
+        x: array of shape (N, D)
+        y: array of shape (M, D)
+
+    Returns:
+        Array of shape (N, M)
+    """
+    assert x.shape[1] == y.shape[1], "Input vectors must have the same dimensionality"
+    # Use broadcasting to compute pairwise differences
+    expanded_x = mx.expand_dims(x, 1)
+    distances = mx.linalg.norm(expanded_x - y, axis=-1)
+
+    return distances
